requirements.txt
ansimarkup==2.2.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
click-default-group==1.2.4
click-repl==0.3.0
colorama==0.4.6
dict2xml==1.7.8
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
more-itertools==10.8.0
prompt_toolkit==3.0.52
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
urllib3==2.6.3
wcwidth==0.6.0
